Fix Linux build: cast TIOCGWINSZ to UInt for ioctl
On Glibc the TIOCGWINSZ macro imports as Int32 while ioctl expects
UInt, so terminalColumns() failed to compile on the Linux CI runner.
Darwin already imports it as UInt, so the cast is a no-op there.

+++ MaiCore/Sources/mai/TerminalLineEditor.swift
@@ -283,7 +283,7 @@ final class TerminalLineEditor {
 
   static func terminalColumns() -> Int {
     var size = winsize()
-    guard ioctl(STDOUT_FILENO, TIOCGWINSZ, &size) == 0, size.ws_col > 0 else { return 80 }
+    guard ioctl(STDOUT_FILENO, UInt(TIOCGWINSZ), &size) == 0, size.ws_col > 0 else { return 80 }
     return Int(size.ws_col)
   }
